test: LottoPurchaseAmountTest 작성
